[STEP3] Main 추가 LottoController 클래스로 수정
