+++ src/main/kotlin/lotto/Main.kt
@@ -1,0 +1,7 @@
+package lotto
+
+import lotto.controller.LottoGameController
+
+fun main() {
+    LottoGameController.play()
+}
