--- src/main/kotlin/lotto/view/LottoController.kt
@@ -1,13 +1,0 @@
-package lotto.view
-
-import lotto.domain.LottoShop
-
-fun main() {
-    val money = LottoInputView.inputMoney()
-    val lottos = LottoShop.sellByMoney(money)
-    LottoOutputView.printLottos(lottos)
-
-    val winningNumbers = LottoInputView.inputWinningLottoNumbers()
-    val lottosResult = lottos.calculateResults(winningNumbers)
-    LottoOutputView.printLottoResults(lottosResult)
-}
